# Unroll loops for additional performance