--- a/kuznyechik/Cargo.toml
+++ b/kuznyechik/Cargo.toml
@@ -17,3 +17,8 @@ opaque-debug = "0.2"
 
 [dev-dependencies]
 block-cipher = { version = "0.7", features = ["dev"] }
+
+[features]
+# disables loop unrolling, which reduces resulting binary size,
+# but degrades performance in return
+no_unroll = []