On my notebook I get the following results:
```
// with loop unrolling
test decrypt ... bench:         736 ns/iter (+/- 37) = 21 MB/s
test encrypt ... bench:         526 ns/iter (+/- 62) = 30 MB/s
// without loop unrolling
test decrypt ... bench:       1,237 ns/iter (+/- 71) = 12 MB/s
test encrypt ... bench:         996 ns/iter (+/- 165) = 16 MB/s
```

Unrolling can be disabled by enabling the new `no_unroll` feature.